Parsing des elements de configuration de la simulation et chargement des pannes

+++ include/core/Systeme.h
@@ -34,7 +34,7 @@ private:
 
 public:
 	Systeme(double cap);
-	Systeme(double cap, double dur , double consomation);
+	Systeme(double cap, int dur , double consomation);
 	~Systeme();
 
 	//Getters
